Reorganize platform specific code
